+++ download.sh
@@ -2,7 +2,7 @@
 
 # Set version to latest unless set by user
 if [ -z "$VERSION" ]; then
-  VERSION="2.10.2"
+  VERSION="2.10.3"
 fi
 
 echo "Downloading version ${VERSION}..."